moved session cache into own source files
