+++ lib/vtls/vtls.c
@@ -55,6 +55,7 @@
 
 #include "vtls.h" /* generic SSL protos etc */
 #include "vtls_int.h"
+#include "spool.h"
 
 #include "openssl.h"        /* OpenSSL versions */
 #include "gtls.h"           /* GnuTLS versions */
@@ -513,481 +514,6 @@ ssl_connect_nonblocking(struct Curl_cfilter *cf, struct Curl_easy *data,
 
   /* mark this is being ssl requested from here on. */
   return Curl_ssl->connect_nonblocking(cf, data, done);
-}
-
-/*
- * Lock shared SSL session data
- */
-void Curl_ssl_sessionid_lock(struct Curl_easy *data)
-{
-  if(SSLSESSION_SHARED(data))
-    Curl_share_lock(data, CURL_LOCK_DATA_SSL_SESSION, CURL_LOCK_ACCESS_SINGLE);
-}
-
-/*
- * Unlock shared SSL session data
- */
-void Curl_ssl_sessionid_unlock(struct Curl_easy *data)
-{
-  if(SSLSESSION_SHARED(data))
-    Curl_share_unlock(data, CURL_LOCK_DATA_SSL_SESSION);
-}
-
-static CURLcode cf_ssl_conn_hash_add_path(struct dynbuf *buf,
-                                          const char *name,
-                                          char *path)
-{
-  if(path && path[0]) {
-    /* We try to add absolute paths, so that the session key can stay
-     * valid when used in another process with different CWD. However,
-     * when a path does not exist, this does not work. Then, we add
-     * the path as is. */
-#ifdef _WIN32
-    char abspath[_MAX_PATH];
-    if(_fullpath(abspath, path, _MAX_PATH))
-      return Curl_dyn_addf(buf, ":%s-%s", name, abspath);
-#else
-    if(path[0] != '/') {
-      char *abspath = realpath(path, NULL);
-      if(abspath) {
-        CURLcode r = Curl_dyn_addf(buf, ":%s-%s", name, abspath);
-        (free)(abspath); /* allocated by libc, free without memdebug */
-        return r;
-      }
-    }
-#endif
-    return Curl_dyn_addf(buf, ":%s-%s", name, path);
-  }
-  return CURLE_OK;
-}
-
-static CURLcode cf_ssl_conn_hash_add_hash(struct dynbuf *buf,
-                                          const char *name,
-                                          struct curl_blob *blob)
-{
-  CURLcode r = CURLE_OK;
-  if(blob && blob->len) {
-    unsigned char hash[CURL_SHA256_DIGEST_LENGTH];
-    size_t i;
-
-    r = Curl_dyn_addf(buf, ":%s-", name);
-    if(r)
-      goto out;
-    r = Curl_sha256it(hash, blob->data, blob->len);
-    if(r)
-      goto out;
-    for(i = 0; i < CURL_SHA256_DIGEST_LENGTH; ++i) {
-      r = Curl_dyn_addf(buf, "%02x", hash[i]);
-      if(r)
-        goto out;
-    }
-  }
-out:
-  return r;
-}
-
-CURLcode Curl_ssl_conn_hash_make(struct Curl_cfilter *cf,
-                                 const struct ssl_peer *peer,
-                                 char **phash)
-{
-  struct ssl_primary_config *ssl = Curl_ssl_cf_get_primary_config(cf);
-  struct dynbuf buf;
-  unsigned char hash[CURL_SHA256_DIGEST_LENGTH];
-  size_t i;
-  CURLcode r;
-
-  *phash = NULL;
-  Curl_dyn_init(&buf, 10 * 1024);
-
-  r = Curl_dyn_add(&buf, "SESS");
-  if(r)
-    goto out;
-
-  if(ssl->clientcert && ssl->clientcert[0]) {
-    r = Curl_dyn_add(&buf, "-CCERT");
-    if(r)
-      goto out;
-  }
-#ifdef USE_TLS_SRP
-  if(ssl->username || ssl->password) {
-    r = Curl_dyn_add(&buf, "-SRP");
-    if(r)
-      goto out;
-  }
-#endif
-
-  r = Curl_dyn_addf(&buf, ":%s:%d", peer->hostname, peer->port);
-  if(r)
-    goto out;
-
-  switch(peer->transport) {
-  case TRNSPRT_TCP:
-    break;
-  case TRNSPRT_UDP:
-    r = Curl_dyn_add(&buf, ":UDP");
-    break;
-  case TRNSPRT_QUIC:
-    r = Curl_dyn_add(&buf, ":QUIC");
-    break;
-  case TRNSPRT_UNIX:
-    r = Curl_dyn_add(&buf, ":UNIX");
-    break;
-  default:
-    r = Curl_dyn_addf(&buf, ":TRNSPRT-%d", peer->transport);
-    break;
-  }
-  if(r)
-    goto out;
-
-  if(!ssl->verifypeer) {
-    r = Curl_dyn_add(&buf, ":NO-VRFY-PEER");
-    if(r)
-      goto out;
-  }
-  if(!ssl->verifyhost) {
-    r = Curl_dyn_add(&buf, ":NO-VRFY-HOST");
-    if(r)
-      goto out;
-  }
-  if(ssl->verifystatus) {
-    r = Curl_dyn_add(&buf, ":VRFY-OCSP");
-    if(r)
-      goto out;
-  }
-  if(!ssl->verifypeer || !ssl->verifyhost) {
-    if(cf->conn->bits.conn_to_host) {
-      r = Curl_dyn_addf(&buf, ":CHOST-%s", cf->conn->conn_to_host.name);
-      if(r)
-        goto out;
-    }
-    if(cf->conn->bits.conn_to_port) {
-      r = Curl_dyn_addf(&buf, ":CPORT-%d", cf->conn->conn_to_port);
-      if(r)
-        goto out;
-    }
-  }
-
-  if(ssl->version || ssl->version_max) {
-    r = Curl_dyn_addf(&buf, ":TLSVER-%d-%d", ssl->version, ssl->version_max);
-    if(r)
-      goto out;
-  }
-  if(ssl->ssl_options) {
-    r = Curl_dyn_addf(&buf, ":TLSOPT-%x", ssl->ssl_options);
-    if(r)
-      goto out;
-  }
-  if(ssl->cipher_list) {
-    r = Curl_dyn_addf(&buf, ":CIPHER-%s", ssl->cipher_list);
-    if(r)
-      goto out;
-  }
-  if(ssl->cipher_list13) {
-    r = Curl_dyn_addf(&buf, ":CIPHER13-%s", ssl->cipher_list13);
-    if(r)
-      goto out;
-  }
-  if(ssl->curves) {
-    r = Curl_dyn_addf(&buf, ":CURVES-%s", ssl->curves);
-    if(r)
-      goto out;
-  }
-  r = cf_ssl_conn_hash_add_path(&buf, "CA", ssl->CAfile);
-  if(r)
-    goto out;
-  r = cf_ssl_conn_hash_add_path(&buf, "CApath", ssl->CApath);
-  if(r)
-    goto out;
-  r = cf_ssl_conn_hash_add_path(&buf, "CRL", ssl->CRLfile);
-  if(r)
-    goto out;
-  r = cf_ssl_conn_hash_add_path(&buf, "Issuer", ssl->issuercert);
-  if(r)
-    goto out;
-  if(ssl->pinned_key && ssl->pinned_key[0]) {
-    r = Curl_dyn_addf(&buf, ":Pinned-%s", ssl->pinned_key);
-    if(r)
-      goto out;
-  }
-  if(ssl->cert_blob) {
-    r = cf_ssl_conn_hash_add_hash(&buf, "CertBlob", ssl->cert_blob);
-    if(r)
-      goto out;
-  }
-  if(ssl->ca_info_blob) {
-    r = cf_ssl_conn_hash_add_hash(&buf, "CAInfoBlob", ssl->ca_info_blob);
-    if(r)
-      goto out;
-  }
-  if(ssl->issuercert_blob) {
-    r = cf_ssl_conn_hash_add_hash(&buf, "IssuerBlob", ssl->issuercert_blob);
-  }
-
-  r = Curl_sha256it(hash, (unsigned char *)Curl_dyn_ptr(&buf),
-                    Curl_dyn_len(&buf));
-  if(r)
-    goto out;
-  Curl_dyn_reset(&buf);
-  for(i = 0; i < CURL_SHA256_DIGEST_LENGTH; ++i) {
-    r = Curl_dyn_addf(&buf, "%02x", hash[i]);
-    if(r)
-      goto out;
-  }
-
-out:
-  if(!r) {
-    *phash = Curl_dyn_strdup(&buf);
-    if(!*phash)
-      r = CURLE_OUT_OF_MEMORY;
-  }
-  Curl_dyn_free(&buf);
-  return r;
-}
-
-static bool cf_ssl_session_match_auth(struct Curl_ssl_session *entry,
-                                      struct ssl_primary_config *conn_config)
-{
-  if(!Curl_safecmp(entry->clientcert, conn_config->clientcert))
-    return FALSE;
-#ifdef USE_TLS_SRP
-   if(Curl_timestrcmp(entry->srp_username, conn_config->username) ||
-      Curl_timestrcmp(entry->srp_password, conn_config->password))
-     return FALSE;
-#endif
-  return TRUE;
-}
-
-bool Curl_ssl_get_session(struct Curl_cfilter *cf,
-                          struct Curl_easy *data,
-                          const char *ssl_conn_hash,
-                          void **session,
-                          size_t *session_len, /* set 0 if unknown */
-                          char **palpn)
-{
-  struct ssl_primary_config *conn_config = Curl_ssl_cf_get_primary_config(cf);
-  struct ssl_config_data *ssl_config = Curl_ssl_cf_get_config(cf, data);
-  struct Curl_ssl_session *entry;
-  long *general_age;
-  size_t i;
-
-  *session = NULL;
-  *session_len = 0;
-  if(!ssl_config)
-    return FALSE;
-
-  DEBUGASSERT(ssl_config->primary.cache_session);
-  if(!ssl_config->primary.cache_session || !data->state.session)
-    return FALSE;
-
-  /* Lock if shared */
-  if(SSLSESSION_SHARED(data))
-    general_age = &data->share->sessionage;
-  else
-    general_age = &data->state.sessionage;
-
-  for(i = 0; i < data->set.general_ssl.max_ssl_sessions; i++) {
-    entry = &data->state.session[i];
-    if(!entry->ssl_conn_hash)  /* unused entry */
-      continue;
-    if(strcasecompare(ssl_conn_hash, entry->ssl_conn_hash) &&
-       cf_ssl_session_match_auth(entry, conn_config)) {
-      /* yes, we have a cached session for this! */
-      (*general_age)++;          /* increase general age */
-      entry->age = *general_age; /* set this as used in this age */
-      *session = entry->session;
-      if(session_len)
-        *session_len = entry->session_len;
-      if(palpn)
-        *palpn = entry->alpn;
-      break;
-    }
-  }
-
-  CURL_TRC_CF(data, cf, "%s cached session for %s",
-              *session ? "Found" : "No", ssl_conn_hash);
-  return !!*session;
-}
-
-/*
- * Kill a single session ID entry in the cache.
- */
-void Curl_ssl_kill_session(struct Curl_ssl_session *entry)
-{
-  if(entry->ssl_conn_hash) {
-    /* free the ID the SSL-layer specific way */
-    entry->sessionid_free(entry->session, entry->session_len);
-
-    entry->session = NULL;
-    entry->sessionid_free = NULL;
-    entry->age = 0; /* fresh */
-
-    Curl_safefree(entry->alpn);
-    Curl_safefree(entry->clientcert);
-#ifdef USE_TLS_SRP
-    Curl_safefree(entry->srp_username);
-    Curl_safefree(entry->srp_password);
-#endif
-    Curl_safefree(entry->ssl_conn_hash);
-  }
-}
-
-/*
- * Delete the given session from the cache using its key.
- */
-void Curl_ssl_del_session(struct Curl_easy *data, const char *ssl_conn_hash)
-{
-  size_t i;
-
-  for(i = 0; i < data->set.general_ssl.max_ssl_sessions; i++) {
-    struct Curl_ssl_session *entry = &data->state.session[i];
-
-    if(entry->ssl_conn_hash &&
-       strcasecompare(ssl_conn_hash, entry->ssl_conn_hash)) {
-      Curl_ssl_kill_session(entry);
-      break;
-    }
-  }
-}
-
-CURLcode Curl_ssl_add_session(struct Curl_cfilter *cf,
-                              struct Curl_easy *data,
-                              const char *ssl_conn_hash,
-                              const struct ssl_peer *peer,
-                              void *session,
-                              size_t session_len,
-                              Curl_ssl_sessionid_dtor *sessionid_free_cb,
-                              const char *alpn)
-{
-  struct ssl_config_data *ssl_config = Curl_ssl_cf_get_config(cf, data);
-  struct ssl_primary_config *conn_config = Curl_ssl_cf_get_primary_config(cf);
-  size_t i;
-  struct Curl_ssl_session *entry;
-  long oldest_age;
-  char *clone_skey = NULL;
-  char *clone_alpn = NULL;
-  char *clone_clientcert = NULL;
-#ifdef USE_TLS_SRP
-  char *clone_srp_username = NULL;
-  char *clone_srp_password = NULL;
-#endif
-  long *general_age;
-  void *old_session;
-  size_t old_session_len;
-  CURLcode result = CURLE_OUT_OF_MEMORY;
-
-  DEBUGASSERT(session);
-  DEBUGASSERT(sessionid_free_cb);
-
-  if(!data->state.session) {
-    sessionid_free_cb(session, session_len);
-    return CURLE_OK;
-  }
-
-  if(Curl_ssl_get_session(cf, data, ssl_conn_hash,
-                          &old_session, &old_session_len, NULL)) {
-    if((old_session_len == session_len) &&
-       ((old_session == session) ||
-        (session_len && !memcmp(old_session, session, session_len)))) {
-      /* the very same */
-      sessionid_free_cb(session, session_len);
-      return CURLE_OK;
-    }
-    Curl_ssl_del_session(data, ssl_conn_hash);
-  }
-
-  entry = &data->state.session[0];
-  oldest_age = data->state.session[0].age; /* zero if unused */
-  DEBUGASSERT(ssl_config->primary.cache_session);
-  (void)ssl_config;
-
-  clone_skey = strdup(ssl_conn_hash);
-  if(!clone_skey)
-    goto out;
-
-  if(conn_config->clientcert) {
-    clone_clientcert = strdup(conn_config->clientcert);
-    if(!clone_clientcert)
-      goto out;
-  }
-
-#ifdef USE_TLS_SRP
-  if(conn_config->username) {
-    clone_srp_username = strdup(conn_config->username);
-    if(!clone_srp_username)
-      goto out;
-  }
-  if(conn_config->password) {
-    clone_srp_password = strdup(conn_config->password);
-    if(!clone_srp_password)
-      goto out;
-  }
-#endif
-
-  clone_alpn = alpn ? strdup(alpn) : NULL;
-  if(alpn && !clone_alpn)
-    goto out;
-
-  /* Now we should add the session ID and the hostname to the cache, (remove
-     the oldest if necessary) */
-
-  /* If using shared SSL session, lock! */
-  if(SSLSESSION_SHARED(data)) {
-    general_age = &data->share->sessionage;
-  }
-  else {
-    general_age = &data->state.sessionage;
-  }
-
-  /* find an empty slot for us, or find the oldest */
-  for(i = 1; (i < data->set.general_ssl.max_ssl_sessions) &&
-        data->state.session[i].ssl_conn_hash; i++) {
-    if(data->state.session[i].age < oldest_age) {
-      oldest_age = data->state.session[i].age;
-      entry = &data->state.session[i];
-    }
-  }
-  if(i == data->set.general_ssl.max_ssl_sessions)
-    /* cache is full, we must "kill" the oldest entry! */
-    Curl_ssl_kill_session(entry);
-  else
-    entry = &data->state.session[i]; /* use this slot */
-
-  /* now init the session struct wisely */
-  DEBUGASSERT(!entry->ssl_conn_hash);
-  entry->ssl_conn_hash = clone_skey;
-  clone_skey = NULL;
-  entry->session = session;
-  entry->session_len = session_len;
-  entry->sessionid_free = sessionid_free_cb;
-  entry->age = *general_age;    /* set current age */
-  /* free it if there is one already present */
-  DEBUGASSERT(!entry->clientcert);
-  entry->clientcert = clone_clientcert;
-  clone_clientcert = NULL;
-  entry->alpn = clone_alpn;
-  clone_alpn = NULL;
-#ifdef USE_TLS_SRP
-  DEBUGASSERT(!entry->srp_username);
-  entry->srp_username = clone_srp_username;
-  clone_srp_username = NULL;
-  DEBUGASSERT(!entry->srp_password);
-  entry->srp_password = clone_srp_password;
-  clone_srp_password = NULL;
-#endif
-  result = CURLE_OK;
-
-out:
-  free(clone_skey);
-  free(clone_alpn);
-  if(result) {
-    failf(data, "Failed to add SSL Session to cache for %s", ssl_conn_hash);
-    sessionid_free_cb(session, session_len);
-    return result;
-  }
-  CURL_TRC_CF(data, cf, "Added SSL Session to cache for %s:%d [%s]",
-              peer->hostname, peer->port,
-              Curl_ssl_cf_is_proxy(cf) ? "PROXY" : "server");
-  return CURLE_OK;
 }
 
 CURLcode Curl_ssl_get_channel_binding(struct Curl_easy *data, int sockindex,
