+++ lib/vtls/vtls.h
@@ -155,43 +155,6 @@ CURLcode Curl_ssl_push_certinfo(struct Curl_easy *data, int certnum,
 
 /* Functions to be used by SSL library adaptation functions */
 
-/* Lock session cache mutex.
- * Call this before calling other Curl_ssl_*session* functions
- * Caller should unlock this mutex as soon as possible, as it may block
- * other SSL connection from making progress.
- * The purpose of explicitly locking SSL session cache data is to allow
- * individual SSL engines to manage session lifetime in their specific way.
- */
-void Curl_ssl_sessionid_lock(struct Curl_easy *data);
-
-/* Unlock session cache mutex */
-void Curl_ssl_sessionid_unlock(struct Curl_easy *data);
-
-/* create a hash of the ssl connection paramter
- * Create a hash of printable chars for storage of TLS sessions suitable
- * for the given connection filter and peer. The key will reflect the
- * SSL config for the filter's connection (ssl verssions/options/ciphers etc.).
- * Config options using relative paths will be converted to absolute
- * ones. config options involving BLOBs will add the SHA256 hash of the
- * BLOB. In such configurations the key is not guarantueed to be unique,
- * but collisions are highly unlikely since they would involve the same
- * peer and matching other config options.
- * @param cf      the connection filter wanting to use it
- * @param peer    the peer the filter wants to talk to
- * @param phash   on successfull return, the hash generated
- */
-CURLcode Curl_ssl_conn_hash_make(struct Curl_cfilter *cf,
-                                 const struct ssl_peer *peer,
-                                 char **phash);
-
-/* Kill a single session ID entry in the cache
- * Sessionid mutex must be locked (see Curl_ssl_sessionid_lock).
- * This will call engine-specific curlssl_session_free function, which must
- * take sessionid object ownership from sessionid cache
- * (e.g. decrement refcount).
- */
-void Curl_ssl_kill_session(struct Curl_ssl_session *session);
-
 /* get N random bytes into the buffer */
 CURLcode Curl_ssl_random(struct Curl_easy *data, unsigned char *buffer,
                          size_t length);
